+++ CHANGELOG.md
@@ -12,6 +12,10 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 **CHANGES**
 - Ubuntu 20.04 is no longer supported.
 
+**BUG FIXES**
+- Fix a bug in the installation of ARM Performance Library that was causing the download of GCC dependencies from 
+  gcc website rather than the ParallelCluster public bucket.
+
 3.13.0
 ------
 **ENHANCEMENTS**